updated customer list when checkedout button is clicked
